feat: Refactor backend code and add user progress tracking in speaking assessment
- Refactored common backend logic for improved maintainability.
- Added user progress update functionality in the speaking assessment module.
- Updated user schema to support new progress tracking fields.
