+++ backend/src/utils/assessmenAnalysisHelper/generateAudioDuration.js
@@ -1,0 +1,14 @@
+const getAudioDuration = async (audioPath) => {
+  try {
+    // Parse the audio file
+    const metadata = await parseFile(audioPath);
+    const durationInSeconds = metadata.format.duration; // Duration in seconds
+    console.log(`Audio Duration: ${durationInSeconds} seconds`);
+    return durationInSeconds;
+  } catch (error) {
+    console.error("Error reading audio file:", error);
+    throw new Error("Failed to retrieve audio duration");
+  }
+};
+
+export { getAudioDuration as getAudioDuration};
